% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.prep.R
\name{data.prep}
\alias{data.prep}
\title{data.prep}
\usage{
data.prep(M, label)
}
\arguments{
\item{M}{A m \times n matrix containing SNP values 0,1,2 of each of the n
alleles of m sequences.}

\item{label}{A integer vector of population group label for each individuals.
Indexing of label is the same as row index of M.}
}
\value{
List of SNP matrix, each includes all individuals from a population.
}
\description{
Preapre data for Fst calculation. Parse SNP dataset by population.
}
\examples{

data.prep(M,label)

}
