% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fst.R
\name{FST}
\alias{FST}
\title{the Fixation Index}
\usage{
FST(pop1, pop2, method = "NUll")
}
\arguments{
\item{pop1}{A m x n matrix containing SNP values 0,1,2 of each of the n
alleles of m sequences from population 1.}

\item{pop2}{A m x n matrix containing SNP values 0,1,2 of each of the n
alleles of m sequences from population 2.}

\item{method}{a character string indicating format of output. This must be
'average'(default) or 'list'.}
}
\value{
the average FST value between the two populations.
}
\description{
A measure of population differentiation due to genetic structure.
In this function, Fst is estimated from one of genetic polymorphism data,
single-nucleotide polymorphisms(SNP). Calculation implemented by Hudson method
in Hudson(1992).
}
\examples{

SNP_example=sim_SNP(method="SNP")
label=sim_SNP(method="label")
SNP=data.prep(SNP_example,label)

FST(SNP$'1',SNP$'3')
FST(SNP$'2',SNP$'3',method='average')
FST(SNP$'2',SNP$'3',method='list')

}
\references{
Bhatia, G., Patterson, N., Sankararaman, S., and Price, A.L. (2013).
Hudson, R.R., Slatkin, M., and Maddison, W.P. (1992). Estimation of levels of
gene flow from DNA sequence data. Genetics 132, 583-589.
}
