% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{linear_regression}
\alias{linear_regression}
\title{Linear Regression}
\usage{
linear_regression(X, Y, intercept = T)
}
\arguments{
\item{X}{The predictors/independent variables of a model}

\item{Y}{The response/dependent variables of a model}

\item{intercept}{The default of intercept is True}
}
\value{
The function returns some significant statistics of a model
}
\description{
The function will fit a linear regression model which comes from the predictors X and response Y in a dataset.
}
\examples{
model_car <- linear_regression(X = dt[,c(2,3)],Y = dt$mpg)

}
