% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_SNP.R
\name{sim_SNP}
\alias{sim_SNP}
\title{data.simulation}
\usage{
sim_SNP(method = "NULL")
}
\arguments{
\item{method}{NULL, SNP, or Label. Randomly generate SNP dataset and population grouping for tutorial.}
}
\value{
List of SNP matrix and Label. Or matrix of SNP values. Or vector of population grouping labels.
}
\description{
simulate a SNP dataset.
}
\examples{

sim_SNP()

sim_SNP(method="SNP")

sim_SNP(method="label")



}
