% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fst.R
\name{FST}
\alias{FST}
\alias{the}
\alias{Fixation}
\alias{index}
\title{FST}
\usage{
FST(pop1, pop2, method = "NUll")
}
\arguments{
\item{pop1}{A m \times n matrix containing SNP values 0,1,2 of each of the n
alleles of m sequences from population 1.}

\item{pop2}{A m \times n matrix containing SNP values 0,1,2 of each of the n
alleles of m sequences from population 2.}

\item{method}{a character string indicating format of output. This must be
'average'(default) or 'list'.}
}
\value{
the average FST value between the two populations.
}
\description{
A measure of population differentiation due to genetic structure.
In this function, Fst is estimated from genetic polymorphism data, such as
single-nucleotide polymorphisms(SNP). Calculation implemented by Hudson method
in Hudson(1992).
}
\examples{
val=FST(pop1,pop2)
#The average Fst value between population 1 and population 2 is
print(val)
.

val=FST(pop1,pop2,method='average')
list=FST(pop1,pop2,method='list')
The average Fst value between population 1 and population 2 is
print(val)
.Fst value for each allele is
print(list)
.

}
\references{
Bhatia, G., Patterson, N., Sankararaman, S., and Price, A.L. (2013).
Hudson, R.R., Slatkin, M., and Maddison, W.P. (1992). Estimation of levels of
gene flow from DNA sequence data. Genetics 132, 583-589.
}
